* origin/master: (74 commits)
  Update doc submodule [nomail] [skip ci]
  Add STORAGE_ prefixes for backends and serializers
  Add versioning to JSON serializer
  Remove unnecessary includes in Val.h
  Move byte_buffer types from cluster and storage into util
  Remove unnecessary <array> and <memory> includes from util.h
  Mark storage classes as final where appropriate
  Add JSON storage serializer, use with existing backends/tests
  Make ValFromJSON return zeek::expected instead of a variant
  Ground work for pluggable storage serializers
  Bump zeekjs to v0.17.0
  Update bifcl submodule [nomail]
  Prevent event timestamps set to future
  Bump Spicy.
  cluster/Backend: Add name and lookup component tag
  cluster/Event: Hide members behind accessors
  cluster/PublishEvent:: Make event non-const
  broker/Manager: Re-use broker serializer for conversion
  EventMgr: Add Dispatch() with handler and args
  plugin/Manager: Fix MetaHookPre and MetaHookPost using HOOK_CALL_FUNCTION
  ...

+++ src/packet_analysis/Analyzer.h
@@ -108,7 +108,7 @@ public:
      * @param child The analyzer that will be called for the new protocol during
      * forwarding.
      */
-    void RegisterProtocol(uint32_t identifier, AnalyzerPtr child);
+    void RegisterProtocol(uint64_t identifier, AnalyzerPtr child);
 
     /**
      * Registers an analyzer to use for protocol detection if identifier
@@ -202,7 +202,7 @@ protected:
      * @return The analyzer registered for the given identifier. Returns a
      * nullptr if no analyzer is registered.
      */
-    const AnalyzerPtr& Lookup(uint32_t identifier) const;
+    const AnalyzerPtr& Lookup(uint64_t identifier) const;
 
     /**
      * Returns an analyzer based on a script-land definition.
@@ -238,7 +238,7 @@ protected:
      *
      * @return false if the analysis failed, else true.
      */
-    bool ForwardPacket(size_t len, const uint8_t* data, Packet* packet, uint32_t identifier) const;
+    bool ForwardPacket(size_t len, const uint8_t* data, Packet* packet, uint64_t identifier) const;
 
     /**
      * Triggers default analysis of the encapsulated packet if the default analyzer
@@ -266,7 +266,7 @@ private:
                                       const zeek::Tag& arg_tag);
 
     // Internal helpers to find an appropriate next inner analyzer.
-    const AnalyzerPtr& FindInnerAnalyzer(size_t len, const uint8_t* data, Packet* packet, uint32_t identifier) const;
+    const AnalyzerPtr& FindInnerAnalyzer(size_t len, const uint8_t* data, Packet* packet, uint64_t identifier) const;
     const AnalyzerPtr& FindInnerAnalyzer(size_t len, const uint8_t* data, Packet* packet) const;
     const AnalyzerPtr& DetectInnerAnalyzer(size_t len, const uint8_t* data, Packet* packet) const;
